:white_check_mark: simplify tests folder structure
