+++ src/test/java/com/gildedrose/items/AgedBrieItemTest.java
@@ -1,4 +1,4 @@
-package com.gildedrose.items.items;
+package com.gildedrose.items;
 
 import com.gildedrose.main.Item;
 import org.junit.jupiter.api.MethodOrderer.OrderAnnotation;
@@ -6,9 +6,9 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 
+import static com.gildedrose.helper.TestHelper.testItem;
+import static com.gildedrose.helper.TestHelper.testItemException;
 import static com.gildedrose.item_helpers.ItemName.AGED_BRIE;
-import static com.gildedrose.items.helper.TestHelper.testItem;
-import static com.gildedrose.items.helper.TestHelper.testItemException;
 
 @TestMethodOrder(OrderAnnotation.class)
 class AgedBrieItemTest {
