+++ src/test/java/com/gildedrose/items/LegendaryItemTest.java
@@ -1,4 +1,4 @@
-package com.gildedrose.items.items;
+package com.gildedrose.items;
 
 import com.gildedrose.main.GildedRose;
 import com.gildedrose.main.Item;
@@ -7,11 +7,11 @@ import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestMethodOrder;
 
+import static com.gildedrose.helper.TestHelper.testItem;
+import static com.gildedrose.helper.TestHelper.testItemException;
 import static com.gildedrose.item_helpers.ItemName.LEGENDARY;
 import static com.gildedrose.items.LegendaryItem.LEGENDARY_ITEM_QUALITY;
 import static com.gildedrose.items.LegendaryItem.NOT_LEGENDARY_ITEM_ERROR_MESSAGE;
-import static com.gildedrose.items.helper.TestHelper.testItem;
-import static com.gildedrose.items.helper.TestHelper.testItemException;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
